refactor: inject toolbar script from plugin
