+++ src/runtime/plugin.client.ts
@@ -3,10 +3,10 @@ import { defineNuxtPlugin, refreshNuxtData } from '#app'
 import { PrismicModuleOptions } from '../types'
 
 export default defineNuxtPlugin((nuxtApp) => {
-	const mergedOptions: PrismicModuleOptions = useRuntimeConfig().public.prismic
+	const options: PrismicModuleOptions = useRuntimeConfig().public.prismic
 
 	// Hot reload preview updates
-	if (mergedOptions.preview) {
+	if (options.preview) {
 		window.addEventListener('prismicPreviewUpdate', (event) => {
 			event.preventDefault()
 			refreshNuxtData()
